build: skip packing for plugins from orbit and download packed and signed file instead

+++ org.eclipse.scout.releng.ant/src/org/eclipse/scout/releng/ant/pack200/Pack.java
@@ -4,7 +4,7 @@
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
- * 
+ *
  * Contributors:
  *     BSI Business Systems Integration AG - initial API and implementation
  ******************************************************************************/
@@ -14,7 +14,6 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.util.ArrayList;
 import java.util.jar.JarFile;
 import java.util.jar.Pack200.Packer;
 import java.util.zip.GZIPOutputStream;
@@ -22,39 +21,17 @@ import java.util.zip.GZIPOutputStream;
 import org.apache.tools.ant.BuildException;
 import org.apache.tools.ant.DirectoryScanner;
 import org.apache.tools.ant.Project;
-import org.apache.tools.ant.Task;
 import org.apache.tools.ant.types.FileSet;
+import org.eclipse.scout.releng.ant.util.FileUtility;
 
 /**
  * <h4>Repack</h4>
- * 
+ *
  * @author aho
  * @since 1.1.0 (26.01.2011)
  */
-public class Pack extends Task {
-
-  private File outputDir;
-  private final ArrayList<FileSet> filesets;
+public class Pack extends AbstractPack {
   private boolean gzip = true;
-
-  public Pack() {
-    this.filesets = new ArrayList<FileSet>();
-  }
-
-  /**
-   * @param outputDir
-   *          the outputDir to set
-   */
-  public void setOutputDir(File outputDir) {
-    this.outputDir = outputDir;
-  }
-
-  /**
-   * @return the outputDir
-   */
-  public File getOutputDir() {
-    return outputDir;
-  }
 
   /**
    * @param gzip
@@ -71,10 +48,6 @@ public class Pack extends Task {
     return gzip;
   }
 
-  public void addFileset(FileSet set) {
-    filesets.add(set);
-  }
-
   @Override
   public void execute() throws BuildException {
     validate();
@@ -88,7 +61,7 @@ public class Pack extends Task {
       for (String fileName : ds.getIncludedFiles()) {
         File inputFile = new File(inputDir, fileName);
         try {
-          packFile(inputFile, outputDir.getAbsoluteFile()+File.separator+ fileName+".pack");
+          packFile(inputFile, outputDir.getAbsoluteFile() + File.separator + fileName + ".pack");
         }
         catch (IOException e) {
           throw new BuildException("could not repack file '" + inputFile + "'.", e);
@@ -98,34 +71,38 @@ public class Pack extends Task {
   }
 
   private void packFile(File inputFile, String outputFilename) throws IOException {
-    log("pack200: pack file '"+inputFile.getName()+"'", Project.MSG_INFO);
+    log("pack200: pack file '" + inputFile.getName() + "'", Project.MSG_INFO);
     Packer packer = Pack200Utility.createPacker();
     OutputStream out = null;
-    try {
-      if (isGzip()) {
-        outputFilename += ".gz";
-      }
-      File outputFile = new File(outputFilename);
-      if(!outputFile.exists()){
-        outputFile.getParentFile().mkdirs();
-      }
-      out = new FileOutputStream(outputFile);
-      if(isGzip()){
-        out = new GZIPOutputStream(out);
-      }
-      packer.pack(new JarFile(inputFile), out);
-    }
-    finally {
-      if (out != null) {
-        out.close();
-      }
-    }
-  }
 
-  private void validate() throws BuildException {
-    if (filesets.size() == 0) {
-      throw new BuildException("need to specify a fileset");
+    if (isGzip()) {
+      outputFilename += ".gz";
     }
+    File outputFile = new File(outputFilename);
+    if (!outputFile.exists()) {
+      outputFile.getParentFile().mkdirs();
+    }
+
+    File existingDropin = getExistingDropin(outputFile.getName());
+    if (existingDropin!=null) {
+      log("pack200: skipping file '" + inputFile.getName() + "'", Project.MSG_INFO);
+      FileUtility.copyToDir(existingDropin, outputFile.getParentFile());
+    }
+    else {
+      try {
+        out = new FileOutputStream(outputFile);
+        if (isGzip()) {
+          out = new GZIPOutputStream(out);
+        }
+        packer.pack(new JarFile(inputFile), out);
+      }
+      finally {
+        if (out != null) {
+          out.close();
+        }
+      }
+    }
+
   }
 
 }
